pyproject.toml
[tool.poetry]
name = "ai-case"
version = "0.1.0"
description = ""
authors = ["Meg Patakota"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
openai = "^1.60.2"
python-dotenv = "^1.0.1"
streamlit = "^1.41.1"
jinja2 = "^3.1.5"
matplotlib = "^3.10.0"
seaborn = "^0.13.2"
networkx = "^3.4.2"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
